Rewrite to the new Color System and font updates
