+++ docs/v5/wiki/color.md
@@ -7,9 +7,9 @@ fusion.WHITE
 The color name is always capitalised as per [PEP 8 → Constants](https://peps.python.org/pep-0008/#constants). All colors are defined and named (with capitalised form) as per [Sublime Text → Docs → Color Schemes → Appendix - CSS Colors](https://www.sublimetext.com/docs/color_schemes.html#appendix-css-colors).
 
 ## Custom color
-If you want your own color, you just give your function a tuple argument with RGBA colors, here is an example:
+If you want your own color, you just give your function a Color object as argument with RGBA colors, here is an example:
 ```python
-fusion.draw_rect(window, 100, 100, 400, 400, (255, 255, 255, 0))
+fusion.draw_rect(window, 100, 100, 400, 400, fusion.Color(255, 255, 255, 255))
 ```
 
 ## Hex to rgba
@@ -23,4 +23,4 @@ If you have a HSV color and want to convert it to RGB you can use this function:
 ```python
 fusion.hsv_to_rgb(h, s, v)
 ```
-hsv_to_rgb(hue, sat, val, alpha) -> tuple[int, int, int, int]
+hsv_to_rgb(hue, sat, val, alpha) -> Color
